project1/requirements.txt
category_encoders==2.6.0
click==8.1.7
itables==2.0.0
marimo==0.3.3
numpy==1.25.2
pandas==2.2.1
pydantic==2.6.4
pytest==8.1.1
python-dotenv==1.0.1
scikit_learn==1.4.1.post1
seaborn==0.13.2
setuptools==58.1.0
sidetable==0.9.1
skimpy==0.0.14
summarytools==0.2.3
ydata_profiling==4.6.5
